requirements.txt
streamlit==1.57.0
pandas==2.3.3
plotly==6.7.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.25.0
db-dtypes==1.5.1
google-genai==1.75.0
pyarrow==24.0.0
requests==2.33.1
